Moved old webapi to api/v1 (#837)
